feat: Built the static version of Bookshelves
Created a static version of the app were I have added data such as "bookshelves" and "books" and passed those to child components to render that. Also did the breakdown of the components and created "Bookshelves", "Bookshelf", "Books" and "Book" components.
Each component is getting props from child component and then rendering the UI.